+++ examples/database/task_script.py
@@ -12,6 +12,7 @@ db = tl.db.TensorHub(ip='localhost', port=27017, dbname='temp', project_name='tu
 # load dataset from database
 X_train, y_train, X_val, y_val, X_test, y_test = db.find_top_dataset('mnist')
 
+
 # define the network
 def mlp():
     ni = tl.layers.Input([None, 784], name='input')
@@ -24,14 +25,17 @@ def mlp():
     M = tl.models.Model(inputs=ni, outputs=net)
     return M
 
+
 network = mlp()
 
 # cost and accuracy
 cost = tl.cost.cross_entropy
 
+
 def acc(y, y_):
     correct_prediction = tf.equal(tf.argmax(y, 1), tf.convert_to_tensor(y_, tf.int64))
     return tf.reduce_mean(tf.cast(correct_prediction, tf.float32))
+
 
 # define the optimizer
 train_op = tf.optimizers.Adam(learning_rate=0.0001)
@@ -43,8 +47,17 @@ train_op = tf.optimizers.Adam(learning_rate=0.0001)
 # )
 
 tl.utils.fit(
-    network, train_op=tf.optimizers.Adam(learning_rate=0.0001), cost=tl.cost.cross_entropy, X_train=X_train,
-    y_train=y_train, acc=acc, batch_size=256, n_epoch=20, X_val=X_val, y_val=y_val, eval_train=False,
+    network,
+    train_op=tf.optimizers.Adam(learning_rate=0.0001),
+    cost=tl.cost.cross_entropy,
+    X_train=X_train,
+    y_train=y_train,
+    acc=acc,
+    batch_size=256,
+    n_epoch=20,
+    X_val=X_val,
+    y_val=y_val,
+    eval_train=False,
 )
 
 # evaluation and save result that match the result_key
@@ -55,5 +68,3 @@ test_accuracy = float(test_accuracy)
 db.save_model(network, model_name='mlp', name=str(n_units1) + '-' + str(n_units2), test_accuracy=test_accuracy)
 # in other script, you can load the model as follow
 # net = db.find_model(sess=sess, model_name=str(n_units1)+'-'+str(n_units2)
-
-tf.python.keras.layers.BatchNormalization