+++ CHANGELOG.md
@@ -70,6 +70,7 @@ To release a new version, please update the changelog as followed:
 ## [Unreleased]
 
 ### Added
+- Support nested layer customization (#PR 1015)
 
 ### Changed
 
@@ -83,7 +84,7 @@ To release a new version, please update the changelog as followed:
 
 ### Fixed
 - Fix `tf.models.Model._construct_graph` for list of outputs, e.g. STN case (PR #1010)
-
+- Enable better `in_channels` exception raise. (pR #1015)
 ### Removed
 
 ### Security
@@ -91,7 +92,7 @@ To release a new version, please update the changelog as followed:
 ### Contributors
 
 - @zsdonghao
-- @ChrisWu1997: #1010
+- @ChrisWu1997: #1010 #1015
 
 ## [2.1.0]